Test stdlib with our tests
We can't test our library against the stdlib's tests, because they
aren't cross-compatible, but we can test the stdlib against our library.
This ensures that we can test for parity between the stdlib and our
library.

This required being able to _not_ install our library, which required
moving the source to a subdirectory, which I called src. This follows
patterns from other languages, as well as the cryptography library.