Testing cleanups (#29)
Python 3.3 doesn't build, because tox no longer supports it. To
get around this, we need to use pytest directly for Python 3.3.

Tox-Travis has deprecated the "after all" feature, in favor of
Travis build stages, which are in general a better solution to
the problem addressed. So this also converts to using them.